+++ lib/strong_migrations.rb
@@ -29,7 +29,7 @@ module StrongMigrations
       :target_postgresql_version, :target_mysql_version, :target_mariadb_version,
       :enabled_checks, :lock_timeout, :statement_timeout, :check_down, :target_version,
       :safe_by_default, :target_sql_mode, :lock_timeout_retries, :lock_timeout_retry_delay,
-      :alphabetize_schema, :skip_databases
+      :alphabetize_schema, :skip_databases, :remove_invalid_indexes
     attr_writer :lock_timeout_limit
   end
   self.auto_analyze = false
@@ -41,6 +41,7 @@ module StrongMigrations
   self.check_down = false
   self.alphabetize_schema = false
   self.skip_databases = []
+  self.remove_invalid_indexes = false
 
   # private
   def self.developer_env?