+++ test/safe_by_default_test.rb
@@ -22,7 +22,13 @@ class SafeByDefaultTest < Minitest::Test
       end
     end
 
-    assert_safe AddIndex
+    assert_raises(ActiveRecord::StatementInvalid) do
+      migrate AddIndex
+    end
+
+    StrongMigrations.stub(:remove_invalid_indexes, true) do
+      assert_safe AddIndex
+    end
   end
 
   def test_add_index_extra_arguments