+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added `skip_databases` option
 - Added warning for unsupported adapters
+- Improved output for `db:rollback`, `db:migrate:up`, and `db:migrate:down`
 
 ## 2.0.2 (2024-10-30)