+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `skip_databases` option
 - Added warning for unsupported adapters
 - Improved output for `db:forward`, `db:rollback`, `db:migrate:up`, and `db:migrate:down`
+- Made operations more retriable with `safe_by_default`
 
 ## 2.0.2 (2024-10-30)